Add basic Arch Linux + LXDE environment (#1)
* Adds ticketted-dev environment [WIP]

Switched to Arch Linux from CentOS/7. Seems to work a little better? Still not
quite happy with the order of operations for VirtualBox Guest Additions, but I
will take what I can get at the moment. Boots to LXDE desktop with the vagrant
user. Would like to change in the near future, but that may involve referencing
a configuration script that is excluded from the repository (secrets and all).

* Disassociate from ticketted project

The repository's focus has shifted from specific environments to collecting the
'base' environments - those environments which are then additionally customized
for a particular project.

I feel a Packer pipeline is in the future. How near? Only the future knows.

* Granulates the ArchLinux envs to include DE

In this particular case, I've been testing/looking to use LXDE, so include that
kind of information in the path. If, in the future, a GNOME-based ArchLinux env
is wanted, I would expect a similar pair of files, under an archlinux/gnome
path.