Fixes inability to complete vagrant up after host upgrade
Running a 1.7G upgrade on my host Arch Linux system, due to a Vagrant issue
with it, resulted in a failure to complete the shell provisioner for the
bootstrap script.

Disabling the script provisioner, updating the box (2020-08-06 to 2020-08-12)
and running gave me an instance to start debugging. Manually executing the
commands in the bootstrap script resulted in success (the changes to the script
are the differences in the commands manually executed -- the Xorg-related
packages seemed to be a possible cause, so removing them let the LXDE package
be successful).

During the manual execution of commands, installing the virtualbox-guest-utils
package will actually remove the virtualbox-guest-utils-nox, so there was no
longer a need to remove it prior. Don't know if this had an impact, other than
a small optimization of the script file itself.

Still haven't figured out why, after a reboot invoked by Vagrant, won't yield a
full-screen display (albeit windowed) for the guest operating system, but even
before logging in to the running instance, manually invoking a reboot from the
guest system will result in a restarted instance with the full-screen display.
Just weird, but as there's a workaround (reboot before logging in for initial
usage), I'm not going to spend much cycles on it.